Adjust update repo for ports
Unlike all the other repositories, the ports updates are not
under /ports/$portsarch/ but under /ports/update (e.g. all
architectures under one directory). Special case this exception
for leap (it is "normal" for tumbleweed). The reason for
this is in how the opensuse maintenance engine works (it
can handle only one Update project for each arches) so it
cannot be changed easily.