add system role for XFCE desktop
Having a system role would make the installer live cd easier to set up. The current setup forces the user to go through the "generic desktop", an extra step that is not intuitive for the average user. Also as the Xfce desktop is already included in the DVD media, it would make sense to have it as on option to select from the start.