Default to wicked on generic desktop role
In the past, yast silently ignored this setting and went for wicked
if NetworkManager did not get installed. Since yast2-network 4.1.40,
this is no longer the case: if the role asked for NetworkManager,
yast now selects NM for installation to comply to he request.

The generic desktop role is a minimal X environment, where no
NetworkManager applet is pre-installed by default. So staying with
wicked, as it used to be by accident, is what we want not by definition.